requirements.txt
pandas==2.2.2
mysql-connector-python==8.4.0
python-dotenv==1.0.1
fsspec==2025.5.1
openpyxl==3.1.2
mysql-connector-python==8.4.0
SQLAlchemy==2.0.30
pymysql